+++ lib/shrine/plugins/add_metadata.rb
@@ -3,20 +3,22 @@ class Shrine
     # The metadata plugin provides a convenient method for extracting and
     # adding custom metadata values.
     #
-    #     plugin :metadata
+    #     plugin :add_metadata
     #
-    #     metadata "exif" do |io, context|
+    #     add_metadata "exif" do |io, context|
     #       MiniMagick::Image.new(io.path).exif
     #     end
     #
-    # This can also be used to override existing metadata values.
-    module Metadata
+    # If the result of the block is nil, the metadata value won't be assigned.
+    #
+    # This plugin can also be used to override existing metadata values.
+    module AddMetadata
       def self.configure(uploader)
         uploader.opts[:metadata] = {}
       end
 
       module ClassMethods
-        def metadata(name, &block)
+        def add_metadata(name, &block)
           opts[:metadata][name] = block
         end
       end
@@ -26,7 +28,8 @@ class Shrine
           metadata = super
 
           opts[:metadata].each do |name, block|
-            metadata[name] = instance_exec(io, context, &block)
+            value = instance_exec(io, context, &block)
+            metadata[name] = value unless value.nil?
           end
 
           metadata
@@ -34,6 +37,6 @@ class Shrine
       end
     end
 
-    register_plugin(:metadata, Metadata)
+    register_plugin(:add_metadata, AddMetadata)
   end
 end